Move the templator into its own class
This allows it to be reused. It also has the code for rendering markdown
and opening the result in an editor
